+++ src/hooks/useArrayFieldState.js
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { ArrayFieldStateContext } from '../Context';
+
+function useArrayFieldState() {
+  return useContext(ArrayFieldStateContext);
+}
+
+export { useArrayFieldState };